More intuitive key assignment - incomplete (#101)
* Modified key configuration to show the keyname instead of the keycode on
the Key button. The keycode is now a data item on the button instead of
the text.
It works correctly for existing key assignments. When adding a key,
however, two keydown events are caught and the second is captured as
the key name.

https://github.com/billmania/roboquest_ui/issues/95

* Tested all of:
redefine an existing key
add a key to a widget with keys
add a key to a widget without any keys
remove an old key from a widget

Updated to config file version 6.

https://github.com/billmania/roboquest_ui/issues/95

* Added check for already assigned key to KeyControl. Removed the list of
assigned keycodes.

https://github.com/billmania/roboquest_ui/issues/95