gamepad support (#138)
* Added a helpless gamepad widget. It can be created and positioned.
It uses the RQ standard "configuration" mechanism, but can't be
configured. It includes a button which will be used to enable
and disable the gamepad.

Updated to rq_ui v29 and config format v9.

Also include two files for testing gamepads: gamepadtest.html and
gamepadtest.js.

https://github.com/billmania/roboquest_ui/issues/120

* Implemented basic functionality to detect connect and disconnect
of one gamepad. Gets the lists of buttons and axes. Uses an
interval to poll the gamepad every RQ_PARAMS.POLLPERIOD_MS
(but doesn't yet do anything with the results).

https://github.com/billmania/roboquest_ui/issues/120

* Gamepad can be enabled and disabled. It detects state changes on
the gamepad.

https://github.com/billmania/roboquest_ui/issues/120

* Added a placeholder to index.htm for the gamepad configuration, added
initial scaffolding for building the gamepad configuration matrix when
a gamepad is connected.

https://github.com/billmania/roboquest_ui/issues/120

* Re-factored gamepad confiuration inputs to a table and added a complete
row to the table for each button and each axis.

https://github.com/billmania/roboquest_ui/issues/120

* Reverted config format version to 8.
gamepad new widget configuration enables and disables gamepad correctly.
However, reconfiguring gamepad widget configuration borks the gamepad
state.

https://github.com/billmania/roboquest_ui/issues/120

* Added an example widget configuration for a gamepad widget, which
uses the new list option for widgetConfig.data.
Added the highlight-row CSS class, for showing which gamepad
button or axis is currently selectd. Added logic to Gamepad
to use this highlighting and it sort of works. Still needs
improvement.
Updated RobotComms to use the list of widgetConfig data objects.

https://github.com/billmania/roboquest_ui/issues/120

* The gamepad widget now has a working Enable/Disable button. When the
gamepad is enabled, it will detect activated buttons/axes and log them.

https://github.com/billmania/roboquest_ui/issues/120

* Implement add widget for a gamepad, which includes some validation
of the configuration before saving it.
A fair amount of refactoring in widget_config.js, to improve comments,
code readability, and efficiency.

https://github.com/billmania/roboquest_ui/issues/120

* Added support for more than 10 buttons or axes.

https://github.com/billmania/roboquest_ui/issues/120

* Builds an action map from each gamepad action (button or axis) to
the corresponding data configuration object.

https://github.com/billmania/roboquest_ui/issues/120

* Implemented parsing of attributes and scaling for both topic and
services.

https://github.com/billmania/roboquest_ui/issues/120

* _pollGamepad, when configuringWidget, mostly highlights the
proper configuration row. When not configuringWidget, builds
an Array actions with the values for each configured button/axis,
in preparation for sending it to a gamepad widget valuesHandler
function.

https://github.com/billmania/roboquest_ui/issues/120

* Connected gamepad buttons and axes to topics and services, with run-time
values and constants. Commands are concatenated by topic/service before
publishing.

https://github.com/billmania/roboquest_ui/issues/120

* Can reconfigure an existing gamepad widget. Incorrectly enables gamepad
after saving updated configuration.

https://github.com/billmania/roboquest_ui/issues/120

* Full functionality for gamepad: add, reconfigure, delete, enable,
and disable. Ready for testing.

https://github.com/billmania/roboquest_ui/issues/120

* Fixed a bug where gamepad wasn't enabled when configuring a new
gamepad widget.

https://github.com/billmania/roboquest_ui/issues/120

* All pieces work correctly in a container with Firefox. Chromium
recognizes the gamepad connect event, but isn't getting any
button or axis changes.
A new URL is required: https://ROBOT-IP/.

https://github.com/billmania/roboquest_ui/issues/120

* Added support for Chromium with gamepads.
Made the background image handling logic more robust with Chromium.

https://github.com/billmania/roboquest_ui/issues/120