Reduce the cognitive load of configuring the gamepad (#148)
* Retrieve the list of services and topics and cache them in the Gamepad
object as _servicesList and _topicsList, via the setServicesTopics()
method. The contents of these two Arrays will be used to populate the
destinationName column based on the value of the destinationType
column.

https://github.com/billmania/roboquest_ui/issues/146

* Changing the value of the gamepad destinationType pulldown calls
Gamepad.fillDestinationNamePulldown().

https://github.com/billmania/roboquest_ui/issues/146

* Gamepad.fillDestinationNamePulldown() can use the value of
destinationType to retrieve the sorted list of service or topic names.
This commit only writes the entries to the console.

https://github.com/billmania/roboquest_ui/issues/146

* Added logic to use the services and topics list to create pulldown menus
for the destinationName column in the gamepad configuration dialog.

https://github.com/billmania/roboquest_ui/issues/146

* Added Gamepad.fillInterface() to fill the interface element when the
destinationName element changes.

https://github.com/billmania/roboquest_ui/issues/146

* When the description field of a gamepad configuration row is erased, all
of the other fields on that row are erased, too.

https://github.com/billmania/roboquest_ui/issues/146

* Erasing the description or the destinationType erases the whole row.

https://github.com/billmania/roboquest_ui/issues/146

* Added the ros_interfaces.js module as utility objects to support the
User Interface and the communication with the ROS graph. Added it to
rq_server.js.

https://github.com/billmania/roboquest_ui/issues/146

* Replaced using the services_topics ROS service with the utilities in the
ros_interfaces.js module.

https://github.com/billmania/roboquest_ui/issues/146

* Added ros_interfaces.js to index.htm.
In index.js, call ros.checkMaps() and remove usage of the
services_topics ROS service.
In wGamepad.js, replace dependence on the servicesTopics data
with the ros_interfaces.js objects.
In ros_interfaces.js, set a placeholder attribute for the Empty
ROS interface. Move the getRosTimestamp() utility function to this
module.

These changes maintain functionality in the gamepad configuration
process, for the description, destinationType, destinationName, and
interface columns. There's no logic to assist the attributes column.

https://github.com/billmania/roboquest_ui/issues/146

* gamepad configuration now includes a pulldown menu of attributes.
Also, the color scheme for the pulldown menues are sensible now.

https://github.com/billmania/roboquest_ui/issues/146

* Added capability to validate the gamepad configuration attributes, based
on the interface.

https://github.com/billmania/roboquest_ui/issues/146

* Improved robustness related to validating attributes.

https://github.com/billmania/roboquest_ui/issues/146

* Updated package.xml and src/params.js to v32.

https://github.com/billmania/roboquest_ui/issues/146

* Caught an exception in the NodeJS server raised when a widget config was
incomplete or had an error.

https://github.com/billmania/roboquest_ui/issues/146