Reduced cognitive load for configuring non-gamepad widgets (#151)
* Added a blank option to the widget type menu for configuration, to
detect every widget type change. When the widget type is blank, the
format section of the input form is hidden.
Added the utility function rosConfiguration which is called each time
the widget type is changed.
Added the constant WIDGET_TYPES to validate the type selection.
Added the constant widgetInterface to provide the configuration options
for the non-gamepad widget types. This constant is provides information
similar to src/ros_interfaces.js, but much simpler.
Corrected two bugs in widget_config.js related to undefined widgetType.

https://github.com/billmania/roboquest_ui/issues/149

* Added resetConfigInputs function to clear all of the data input elements
and the widgetType select element, when the dialog is opened to create a
new widget.

https://github.com/billmania/roboquest_ui/issues/149

* Removed the showConfigDefaults variable, because it was not necessary.
Added two arguments to populateWidgetConfigurationDialog, to distinguish
between an 'add' and a 'reconfigure'; and to identify the widgetType
in both scenarios.
Added placeholder functions pickAttributes and dataConfigChange and
removed the rosConfiguration function.
Fixed a bug where configuringWidget wasn't being reset when necessary.

https://github.com/billmania/roboquest_ui/issues/149

* Removed the unused "socket" CSS class from the widget configuration
logic.
Added the CSS class <widgetType>-class to the data elements.
Converted the 'value' widget configuration form from INPUT elements to
SELECT elements.

https://github.com/billmania/roboquest_ui/issues/149

* Added the allWidgetsClass and SELECT elements for the data section of
the widget configuration form in index.htm.
Removed setWidgetConfigDefaults function. Constrained the selection of
configuration elements to only those for the current widget type.
Renamed configureNewWidget HTML ID ot configureNewWidgetForm, for better
specificity.

https://github.com/billmania/roboquest_ui/issues/149

* Adding a new widget and reconfiguring an existing widget correctly
shows the appropriate widget configuration dialog, for all of value,
indicator, button, slider, joystick, and gamepad widgets. It does not
put appropriate values into the data section of the dialog.

https://github.com/billmania/roboquest_ui/issues/149

* When adding a new widget, the data section of the configuration is
initialized using the widgetInterfaces constant and the
setupWidgetDataSection function.

https://github.com/billmania/roboquest_ui/issues/149

* Setup a callback on changes to the SELECT elements in the data
configuration section. Enable the callback to identify the widgetType,
the specific SELECT, and the value.

https://github.com/billmania/roboquest_ui/issues/149

* Added ELEMENT IDs to the data configuration SELECTS, required by the
dataConfigChange function from the previous commit.

https://github.com/billmania/roboquest_ui/issues/149

* Configuration SELECT and INPUT elements can be populated from the
contents of the widgetInterface object. Double-click on the attributes
INPUT element calls the logic-free pickAttributes function.
There is a defect where the widgetType-topic SELECT element has all of
its OPTION elements removed, but still retains the most recent value in
the UI.

https://github.com/billmania/roboquest_ui/issues/149

* Support double-click on the attributes element by opening an attribute
picker dialog for widget attributes. Selected attributes can be appended
to the collection.
Replaced pickAttributes function with combination of showAttributes,
setupTopicAttributeSelect, and setupServiceAttributeSelect.

https://github.com/billmania/roboquest_ui/issues/149

* Add new widgets for all widget types and write them to the configuration
file.
Does not populate with the configuration of an existing widget.

https://github.com/billmania/roboquest_ui/issues/149

* Adding a new widget starts with a reset collection of INPUT elements.
Re-configuring an existing widget fills in the INPUT elements with the
current values. The SELECT elements are not populated.

https://github.com/billmania/roboquest_ui/issues/149

* Examples of how to manipulate SELECTs in FORMs.

https://github.com/billmania/roboquest_ui/issues/149

* Remover jQuery.selectmenu() from #newWidgetType. Configuration menu
is populated correctly. Widgets can be added, moved, and deleted.

https://github.com/billmania/roboquest_ui/issues/149

* Removed jQuery.selectmenu() from association with dataConfigChange.
Logic still does NOT assign a unique widget label when adding a widget.

https://github.com/billmania/roboquest_ui/issues/149

* Reverted the topicDirection configuration element to an INPUT.
Can now add new widgets and re-configure existing widgets. The CHECK
button on the widget attributes configuration dialog does nothing.

https://github.com/billmania/roboquest_ui/issues/149

* Restored all of the Control attributes for configuration.

https://github.com/billmania/roboquest_ui/issues/150

* Version 33.
Fixed defects in assignValue and wSlider related to multiple attributes.

https://github.com/billmania/roboquest_ui/issues/149